+++ scripts/lint.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
 # run various linters
+set -e
 poetry run flake8 --max-line-length 120 yellowbox tests
-poetry run python -c "import sys; sys.exit(sys.version_info < (3,8,0))"
+poetry run python -c "import sys; sys.exit(sys.version_info < (3,9,0))"
 res=$?
 if [ "$res" -eq "0" ]
   then